tests: move another test_utils to tests/tracer/
I think this is the last one now. Also removes the CI job and tox
config.
