+++ tox.ini
@@ -25,7 +25,6 @@ envlist =
     {py27,py35,py36,py37,py38,py39}-profile-minreqs{,-gevent}
     {py27,py35,py36,py37,py38}-integration
     {py27,py35,py36,py37,py38}-ddtracerun
-    {py27,py35,py36,py37,py38}-test_utils
     {py27,py35,py36,py37,py38}-test_logging
 # Integrations environments
     # aiobotocore dropped Python 3.5 support in 0.12
@@ -538,7 +537,6 @@ commands =
     vertica_contrib: pytest {posargs} tests/contrib/vertica/
 # run subsets of the tests for particular library versions
     ddtracerun: pytest {posargs} tests/commands/test_runner.py
-    test_utils: pytest {posargs} tests/contrib/test_utils.py
     test_logging: pytest {posargs} tests/contrib/logging/
     benchmarks: pytest {posargs} tests/benchmark.py
